+++ test/configuration.test.js
@@ -36,6 +36,27 @@ const options = [{
   default: false,
   test: true,
   alt: false
+}, {
+  name: 'useBandwidthFromLocalStorage',
+  default: false,
+  test: true
+}, {
+  name: 'customTagParsers',
+  default: [],
+  test: [{
+    expression: /#PARSER/,
+    customType: 'test',
+    segment: true
+  }]
+}, {
+  name: 'customTagMappers',
+  default: [],
+  test: [{
+    expression: /#MAPPER/,
+    map(line) {
+      return '#FOO';
+    }
+  }]
 }];
 
 const CONFIG_KEYS = Object.keys(Config);
@@ -267,7 +288,7 @@ options.forEach((opt) => {
 
     let hls = this.player.tech_.hls;
 
-    assert.equal(hls.options_[opt.name],
+    assert.deepEqual(hls.options_[opt.name],
                 opt.default,
                 `${opt.name} should be default`);
   });
@@ -300,7 +321,7 @@ options.forEach((opt) => {
 
     let hls = this.player.tech_.hls;
 
-    assert.equal(hls.options_[opt.name],
+    assert.deepEqual(hls.options_[opt.name],
                 opt.test,
                 `${opt.name} should be equal to sourceHandler Option`);
   });
@@ -318,7 +339,7 @@ options.forEach((opt) => {
 
     let hls = this.player.tech_.hls;
 
-    assert.equal(hls.options_[opt.name],
+    assert.deepEqual(hls.options_[opt.name],
                 opt.test,
                 `${opt.name} should be equal to src option`);
   });
@@ -337,7 +358,7 @@ options.forEach((opt) => {
 
     let hls = this.player.tech_.hls;
 
-    assert.equal(hls.options_[opt.name],
+    assert.deepEqual(hls.options_[opt.name],
                 opt.test,
                 `${opt.name} should be equal to sourchHandler option`);
   });
@@ -357,7 +378,7 @@ options.forEach((opt) => {
 
     let hls = this.player.tech_.hls;
 
-    assert.equal(hls.options_[opt.name],
+    assert.deepEqual(hls.options_[opt.name],
                 opt.test,
                 `${opt.name} should be equal to sourchHandler option`);
   });