+++ test/master-playlist-controller.test.js
@@ -67,6 +67,7 @@ QUnit.module('MasterPlaylistController', {
     this.env.restore();
     this.mse.restore();
     videojs.Hls.supportsNativeHls = this.origSupportsNativeHls;
+    window.localStorage.clear();
     videojs.browser = this.oldBrowser;
     this.player.dispose();
   }
@@ -2325,6 +2326,9 @@ QUnit.test('switches off subtitles on subtitle errors', function(assert) {
   assert.equal(textTracks[1].kind, 'subtitles', 'kind is subtitles');
   textTracks[1].mode = 'showing';
 
+  // Wait for VTT segment to be requested
+  this.clock.tick(1);
+
   assert.equal(this.requests.length, 1, 'made a request');
   assert.equal(textTracks[1].mode, 'showing', 'text track still showing');
 
@@ -2340,6 +2344,9 @@ QUnit.test('switches off subtitles on subtitle errors', function(assert) {
 
   // re-enable first text track
   textTracks[1].mode = 'showing';
+
+  // Wait for VTT segment request to be made
+  this.clock.tick(1);
 
   assert.equal(this.requests.length, 1, 'made a request');
   assert.equal(textTracks[1].mode, 'showing', 'text track still showing');
@@ -2398,6 +2405,9 @@ QUnit.test('pauses subtitle segment loader on tech errors', function(assert) {
   assert.notEqual(textTracks[0].kind, 'subtitles', 'kind is not subtitles');
   assert.equal(textTracks[1].kind, 'subtitles', 'kind is subtitles');
   textTracks[1].mode = 'showing';
+
+  // Wait for VTT segment request to be made
+  this.clock.tick(1);
 
   let pauseCount = 0;
 
@@ -2460,6 +2470,9 @@ QUnit.test('disposes subtitle loaders on dispose', function(assert) {
   assert.equal(textTracks[1].kind, 'subtitles', 'kind is subtitles');
   textTracks[1].mode = 'showing';
 
+  // Wait for VTT segment request to be made
+  this.clock.tick(1);
+
   assert.ok(masterPlaylistController.mediaTypes_.SUBTITLES.activePlaylistLoader,
             'has a subtitle playlist loader');
   assert.ok(masterPlaylistController.subtitleSegmentLoader_,
@@ -2506,6 +2519,9 @@ QUnit.test('subtitle segment loader resets on seeks', function(assert) {
   assert.notEqual(textTracks[0].kind, 'subtitles', 'kind is not subtitles');
   assert.equal(textTracks[1].kind, 'subtitles', 'kind is subtitles');
   textTracks[1].mode = 'showing';
+
+  // Wait for VTT segment request to be made
+  this.clock.tick(1);
 
   let resetCount = 0;
   let abortCount = 0;